Promote iter function to payload storage interface
