+++ lib/segment/src/payload_storage/in_memory_payload_storage.rs
@@ -2,7 +2,6 @@ use std::collections::HashMap;
 
 use common::types::PointOffsetType;
 
-use crate::common::operation_error::OperationResult;
 use crate::types::Payload;
 
 /// Same as `SimplePayloadStorage` but without persistence
@@ -15,18 +14,5 @@ pub struct InMemoryPayloadStorage {
 impl InMemoryPayloadStorage {
     pub fn payload_ptr(&self, point_id: PointOffsetType) -> Option<&Payload> {
         self.payload.get(&point_id)
-    }
-
-    pub fn iter<F>(&self, mut callback: F) -> OperationResult<()>
-    where
-        F: FnMut(PointOffsetType, &Payload) -> OperationResult<bool>,
-    {
-        for (key, val) in self.payload.iter() {
-            let do_continue = callback(*key, val)?;
-            if !do_continue {
-                return Ok(());
-            }
-        }
-        Ok(())
     }
 }
