+++ lib/segment/src/payload_storage/in_memory_payload_storage_impl.rs
@@ -73,6 +73,19 @@ impl PayloadStorage for InMemoryPayloadStorage {
     fn flusher(&self) -> Flusher {
         Box::new(|| Ok(()))
     }
+
+    fn iter<F>(&self, mut callback: F) -> OperationResult<()>
+    where
+        F: FnMut(PointOffsetType, &Payload) -> OperationResult<bool>,
+    {
+        for (key, val) in self.payload.iter() {
+            let do_continue = callback(*key, val)?;
+            if !do_continue {
+                return Ok(());
+            }
+        }
+        Ok(())
+    }
 }
 
 #[cfg(test)]
