+++ lib/segment/src/payload_storage/on_disk_payload_storage.rs
@@ -52,22 +52,6 @@ impl OnDiskPayloadStorage {
             .transpose()
             .map_err(OperationError::from)
     }
-
-    pub fn iter<F>(&self, mut callback: F) -> OperationResult<()>
-    where
-        F: FnMut(PointOffsetType, &Payload) -> OperationResult<bool>,
-    {
-        for (key, val) in self.db_wrapper.lock_db().iter()? {
-            let do_continue = callback(
-                serde_cbor::from_slice(&key)?,
-                &serde_cbor::from_slice(&val)?,
-            )?;
-            if !do_continue {
-                return Ok(());
-            }
-        }
-        Ok(())
-    }
 }
 
 impl PayloadStorage for OnDiskPayloadStorage {
@@ -142,5 +126,21 @@ impl PayloadStorage for OnDiskPayloadStorage {
 
     fn flusher(&self) -> Flusher {
         self.db_wrapper.flusher()
+    }
+
+    fn iter<F>(&self, mut callback: F) -> OperationResult<()>
+    where
+        F: FnMut(PointOffsetType, &Payload) -> OperationResult<bool>,
+    {
+        for (key, val) in self.db_wrapper.lock_db().iter()? {
+            let do_continue = callback(
+                serde_cbor::from_slice(&key)?,
+                &serde_cbor::from_slice(&val)?,
+            )?;
+            if !do_continue {
+                return Ok(());
+            }
+        }
+        Ok(())
     }
 }
