+++ lib/segment/src/payload_storage/payload_storage_base.rs
@@ -36,6 +36,14 @@ pub trait PayloadStorage {
 
     /// Return function that forces persistence of current storage state.
     fn flusher(&self) -> Flusher;
+
+    /// Iterate over all stored payload and apply the provided callback.
+    /// Stop iteration if callback returns false or error.
+    ///
+    /// Required for building payload index.
+    fn iter<F>(&self, callback: F) -> OperationResult<()>
+    where
+        F: FnMut(PointOffsetType, &Payload) -> OperationResult<bool>;
 }
 
 pub trait ConditionChecker {
