+++ lib/segment/src/payload_storage/payload_storage_enum.rs
@@ -38,20 +38,6 @@ impl From<OnDiskPayloadStorage> for PayloadStorageEnum {
     }
 }
 
-impl PayloadStorageEnum {
-    pub fn iter<F>(&self, callback: F) -> OperationResult<()>
-    where
-        F: FnMut(PointOffsetType, &Payload) -> OperationResult<bool>,
-    {
-        match self {
-            #[cfg(feature = "testing")]
-            PayloadStorageEnum::InMemoryPayloadStorage(s) => s.iter(callback),
-            PayloadStorageEnum::SimplePayloadStorage(s) => s.iter(callback),
-            PayloadStorageEnum::OnDiskPayloadStorage(s) => s.iter(callback),
-        }
-    }
-}
-
 impl PayloadStorage for PayloadStorageEnum {
     fn overwrite(&mut self, point_id: PointOffsetType, payload: &Payload) -> OperationResult<()> {
         match self {
@@ -127,6 +113,18 @@ impl PayloadStorage for PayloadStorageEnum {
             PayloadStorageEnum::InMemoryPayloadStorage(s) => s.flusher(),
             PayloadStorageEnum::SimplePayloadStorage(s) => s.flusher(),
             PayloadStorageEnum::OnDiskPayloadStorage(s) => s.flusher(),
+        }
+    }
+
+    fn iter<F>(&self, callback: F) -> OperationResult<()>
+    where
+        F: FnMut(PointOffsetType, &Payload) -> OperationResult<bool>,
+    {
+        match self {
+            #[cfg(feature = "testing")]
+            PayloadStorageEnum::InMemoryPayloadStorage(s) => s.iter(callback),
+            PayloadStorageEnum::SimplePayloadStorage(s) => s.iter(callback),
+            PayloadStorageEnum::OnDiskPayloadStorage(s) => s.iter(callback),
         }
     }
 }
